Correções 26/08/2021

+++ lib/data/mappers/recebimento/acordo_mapper.dart
@@ -8,7 +8,7 @@ extension AcordoMapper on AcordoEntity {
     return AcordoEntity(
       numAco: map['NUMACO'],
       unidade: map['UNIDADE'],
-      valTax: map['VALTAX'],
+      valTax: map['VALTAX'].toDouble(),
       totPar: map['TOTPAR'],
       datAco: DateTime.parse(map['DATACO']),
       tipAco: map['TIPACO'],
